@@ -1,8 +1,8 @@
-from .lib import get_schema_cli, apply_patches, generate_schema, process_schema
-
+from .lib import apply_patches, generate_schema, get_schema_cli
 
 if __name__ == "__main__":
     schema = get_schema_cli()
     apply_patches(schema)
     generate_schema(schema, "codegen")
+    print()
     print("All done! ^_^")
